fix(intel): Handle cold reset via physical reset switch
On the Agilex5 platform when cold reset is issued via
CLI application in the OS, it is received in the BL31
via a SMC call and handled accordingly like flush/invalidate
the caches. However, when the cold reset is issued via an
external switch these handlings are missed.
This patch addresses those missed cache operations.

Signed-off-by: Girisha Dengi <girisha.dengi@intel.com>